LayoutTests/inspector/debugger/network-uisourcecode-provider-expected.txt

Tests NetworkUISourceCodeProvider class.


Running: testDocumentResource
Creating resource.
UISourceCode: http://fake.url/1
UISourceCode is editable: false
UISourceCode is content script: false
Mime type: text/html
UISourceCode content: <document resource content>

Running: testScriptResource
Creating resource.
UISourceCode: http://fake.url/2
UISourceCode is editable: true
UISourceCode is content script: false
Mime type: text/javascript
UISourceCode content: <script resource content>

Running: testStylesheetResource
Creating resource.
UISourceCode: http://fake.url/3
Mime type: text/css
UISourceCode content: <stylesheet resource content>
